+++ src/paco/models/events.py
@@ -1,8 +1,12 @@
 from paco.models import schemas
-from paco.models.base import Resource
+from paco.models.base import Resource, Named
 from zope.interface import implementer
 from zope.schema.fieldproperty import FieldProperty
 
+@implementer(schemas.IEventTarget)
+class EventTarget(Named):
+    target = FieldProperty(schemas.IEventTarget['target'])
+    input_json = FieldProperty(schemas.IEventTarget['input_json'])
 
 @implementer(schemas.IEventsRule)
 class EventsRule(Resource):
@@ -10,6 +14,7 @@ class EventsRule(Resource):
     description = FieldProperty(schemas.IEventsRule['description'])
     schedule_expression = FieldProperty(schemas.IEventsRule['schedule_expression'])
     enabled_state = FieldProperty(schemas.IEventsRule['enabled_state'])
+    targets = FieldProperty(schemas.IEventsRule['targets'])
 
     def __init__(self, name, parent):
         super().__init__(name, parent)